Browser Use/requirements.txt
browser-use
python-dotenv
langchain-openai